Important Changes. Read DESCRIPTON
Jason, the move script is now universal in terms of being able to change the variables in other scripts.

Please make not the speed is well speed..

JumpVelocity is how high they jump. As of right now Double Jumping still needs to be implemented. But  as for right now. I just wanted the base of it done so you could easily work on stats with ease.