Added 2 new Spells
ALllllriiiiight,

These were a pain in the ass, still are.

So. Basically the first COMBO has been implemented. However I think Unity was bugging out or something because for god sake The collider works half the time. You'll see if you try it out.

Anyways

Numpad 6. Rock Mine. This ability will shoot a slow moving rock. When it hits an enemy it will explode into tiny rocks into the air. I have it set to two rocks. But that can change anytime.

The one bug this has is, the mini rocks WILL NOT activate even if you throw the Test Dummy into Them. They just wont. I've gone over the code several times and spent 2 hours trying to figure it out. Got no where.

NUMPAD 7

Ground Spiral. This will launch the player into the air. (Now you see the combo ;) )

So this was a weird bug. I think it has something to do with a combat script and the rotation. I will look more into it tomorrow.

BUT for some odd reason. The SAME prefab that was dropped into the screen works AS INTENDED.

However when launched by the combat script. It just DOES NOT WORK.

Other then that. All of it is bug free.

ALSO there is a slight bug with the UI.

The Player 2 mana bar can go past 0. Soooo yea. Have fun! Sorry I couldn't get anymore done... Work called me in early.