Main Menu Start
I have added in the Main menu, and changed the name of the scenes to "Combat Testing" and "Character Select" to keep better track of things. I have added in the options for  Arcade, Versus, Story, Training, Options, and Extras and have set up the 'alpha' format for the main menu.